use registration for updates
This is motivated by the desire to add arbitrary integration schemes.
The challenge with the existing approach is that the time-integration
scheme needed to know about methods implemented in the particles
class, and the particles class needed to know about the available
time-integration schemes so correct methods get implemented. Using
the new registration scheme, the integration scheme only needs to
know about the variables that have been registered for updates, and
the particles class doesn't need to know about the integration
scheme(s).