add crude ghost particle
This adds another step in the time-integration: timestep-setup.
This step uses a particle system sweep. Maybe in the future, there
will be particle system setup and system interaction setup.
Here, system interaction setup is used because ghost particles
need to know about the real particles they're related to.
Properties of ghost particles are updated using the state_update
step and a pointer to the original real particle that the ghost
particle is mirroring.