make ghost boundary more generic
Ghost boundary is defined by a normal unit vector and
a point which is used to define the plane that the
ghost particles are mirrored against. The normal
vector points inside so that ghost particles aren't
accidentally generated inside the domain when a real
particle moves outside the domain. Velocity of the
ghost particle is the calculated from the
projection of the real particle's velocity onto the
normal vector.